+++ FirebaseAdmin/FirebaseAdmin/Auth/FirebaseUserManager.cs
@@ -49,7 +49,12 @@ namespace FirebaseAdmin.Auth
             this.baseUrl = string.Format(IdTooklitUrl, args.ProjectId);
         }
 
-        public static FirebaseUserManager Create(FirebaseApp app)
+        public void Dispose()
+        {
+            this.httpClient.Dispose();
+        }
+
+        internal static FirebaseUserManager Create(FirebaseApp app)
         {
             var args = new FirebaseUserManagerArgs
             {
@@ -68,7 +73,7 @@ namespace FirebaseAdmin.Auth
         /// <param name="cancellationToken">A cancellation token to monitor the asynchronous
         /// operation.</param>
         /// <returns>A record of user with the queried id if one exists.</returns>
-        public async Task<UserRecord> GetUserById(
+        internal async Task<UserRecord> GetUserByIdAsync(
             string uid, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (string.IsNullOrEmpty(uid))
@@ -76,19 +81,60 @@ namespace FirebaseAdmin.Auth
                 throw new ArgumentException("User ID cannot be null or empty.");
             }
 
-            var payload = new Dictionary<string, object>()
+            var query = new UserQuery()
             {
-                { "localId", uid },
+                Field = "localId",
+                Value = uid,
+                Label = "uid",
             };
+            return await this.GetUserAsync(query, cancellationToken);
+        }
 
-            var response = await this.PostAndDeserializeAsync<GetAccountInfoResponse>(
-                "accounts:lookup", payload, cancellationToken).ConfigureAwait(false);
-            if (response == null || response.Users == null || response.Users.Count == 0)
+        /// <summary>
+        /// Gets the user data corresponding to the given email address.
+        /// </summary>
+        /// <param name="email">An email address.</param>
+        /// <param name="cancellationToken">A cancellation token to monitor the asynchronous
+        /// operation.</param>
+        /// <returns>A record of user with the queried email if one exists.</returns>
+        internal async Task<UserRecord> GetUserByEmailAsync(
+            string email, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            if (string.IsNullOrEmpty(email))
             {
-                throw new FirebaseException($"Failed to get user by ID: {uid}");
+                throw new ArgumentException("Email cannot be null or empty.");
             }
 
-            return new UserRecord(response.Users[0]);
+            var query = new UserQuery()
+            {
+                Field = "email",
+                Value = email,
+            };
+            return await this.GetUserAsync(query, cancellationToken);
+        }
+
+        /// <summary>
+        /// Gets the user data corresponding to the given phone number.
+        /// </summary>
+        /// <param name="phoneNumber">A phone number.</param>
+        /// <param name="cancellationToken">A cancellation token to monitor the asynchronous
+        /// operation.</param>
+        /// <returns>A record of user with the queried phone number if one exists.</returns>
+        internal async Task<UserRecord> GetUserByPhoneNumberAsync(
+            string phoneNumber, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            if (string.IsNullOrEmpty(phoneNumber))
+            {
+                throw new ArgumentException("Phone number cannot be null or empty.");
+            }
+
+            var query = new UserQuery()
+            {
+                Field = "phoneNumber",
+                Value = phoneNumber,
+                Label = "phone number",
+            };
+            return await this.GetUserAsync(query, cancellationToken);
         }
 
         /// <summary>
@@ -98,7 +144,7 @@ namespace FirebaseAdmin.Auth
         /// <param name="args">The user account data to be updated.</param>
         /// <param name="cancellationToken">A cancellation token to monitor the asynchronous
         /// operation.</param>
-        public async Task UpdateUserAsync(
+        internal async Task UpdateUserAsync(
             UserArgs args, CancellationToken cancellationToken = default(CancellationToken))
         {
             var payload = args.ToUpdateUserRequest();
@@ -116,7 +162,7 @@ namespace FirebaseAdmin.Auth
         /// <param name="uid">A user ID string.</param>
         /// <param name="cancellationToken">A cancellation token to monitor the asynchronous
         /// operation.</param>
-        public async Task DeleteUserAsync(
+        internal async Task DeleteUserAsync(
             string uid, CancellationToken cancellationToken = default(CancellationToken))
         {
             if (string.IsNullOrEmpty(uid))
@@ -136,9 +182,16 @@ namespace FirebaseAdmin.Auth
             }
         }
 
-        public void Dispose()
+        private async Task<UserRecord> GetUserAsync(UserQuery query, CancellationToken cancellationToken)
         {
-            this.httpClient.Dispose();
+            var response = await this.PostAndDeserializeAsync<GetAccountInfoResponse>(
+                "accounts:lookup", query.Build(), cancellationToken).ConfigureAwait(false);
+            if (response == null || response.Users == null || response.Users.Count == 0)
+            {
+                throw new FirebaseException($"Failed to get user with {query.Description}");
+            }
+
+            return new UserRecord(response.Users[0]);
         }
 
         private async Task<TResult> PostAndDeserializeAsync<TResult>(
@@ -193,6 +246,44 @@ namespace FirebaseAdmin.Auth
             catch (HttpRequestException e)
             {
                 throw new FirebaseException("Error while calling Firebase Auth service", e);
+            }
+        }
+
+        /// <summary>
+        /// Represents a query that can be executed against the Firebase Auth service to retrieve user records.
+        /// A query mainly consists of a <see cref="UserQuery.Field"/> and a <see cref="UserQuery.Value"/> (e.g.
+        /// <c>Field = localId</c> and <c>Value = alice</c>). Additionally, a query may also specify a more
+        /// human-readable <see cref="UserQuery.Label"/> for the field, which will appear on any error messages
+        /// produced by the query.
+        /// </summary>
+        private class UserQuery
+        {
+            internal string Field { get; set; }
+
+            internal string Value { get; set; }
+
+            internal string Label { get; set; }
+
+            internal string Description
+            {
+                get
+                {
+                    var label = this.Label;
+                    if (string.IsNullOrEmpty(label))
+                    {
+                        label = this.Field;
+                    }
+
+                    return $"{label}: {this.Value}";
+                }
+            }
+
+            internal Dictionary<string, object> Build()
+            {
+                return new Dictionary<string, object>()
+                {
+                    { this.Field, new string[] { this.Value } },
+                };
             }
         }
     }